requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
openai==1.59.6
python-dotenv==1.0.1